+++ stacks-node/src/tests/signer/v0/mod.rs
@@ -1541,7 +1541,6 @@ pub fn wait_for_block_rejections_from_signers(
     })?;
     Ok(result)
 }
-
 /// Waits for at least 70% of the provided signers to send an update for a block with the specificed burn block height and parent tenure stacks block height and message version
 pub fn wait_for_state_machine_update(
     timeout_secs: u64,
@@ -1551,24 +1550,35 @@ pub fn wait_for_state_machine_update(
     signer_addresses: &[(StacksAddress, u64)],
 ) -> Result<(), String> {
     wait_for(timeout_secs, || {
-        let mut found_updates = HashSet::new();
+        let mut found_updates: HashSet<StacksAddress> = HashSet::new();
         let stackerdb_events = test_observer::get_stackerdb_chunks();
         for chunk in stackerdb_events
             .into_iter()
             .flat_map(|chunk| chunk.modified_slots)
         {
-            let message = SignerMessage::consensus_deserialize(&mut chunk.data.as_slice())
-                .expect("Failed to deserialize SignerMessage");
+            let Ok(message) = SignerMessage::consensus_deserialize(&mut chunk.data.as_slice())
+            else {
+                continue;
+            };
             let SignerMessage::StateMachineUpdate(update) = message else {
                 continue;
             };
             let Some((address, version)) = signer_addresses
                 .iter()
-                .find(|(addr, _)| chunk.verify(addr).unwrap())
+                .find(|(addr, _)| chunk.verify(addr).unwrap_or(false))
             else {
                 continue;
             };
-            let (burn_block, burn_block_height, current_miner) = match (version, &update.content) {
+
+            let negotiated = update
+                .active_signer_protocol_version
+                .min(update.local_supported_signer_protocol_version);
+            if negotiated != *version {
+                continue;
+            }
+
+            let (burn_block, burn_block_height, current_miner) = match (negotiated, &update.content)
+            {
                 (
                     0,
                     StateMachineUpdateContent::V0 {
@@ -1623,7 +1633,7 @@ pub fn wait_for_state_machine_update(
                 }
             };
             // We only need one update to match our conditions
-            found_updates.insert(address);
+            found_updates.insert(address.clone());
         }
         Ok(found_updates.len() > signer_addresses.len() * 7 / 10)
     })
@@ -1636,24 +1646,34 @@ pub fn wait_for_state_machine_update_by_miner_tenure_id(
     signer_addresses: &[(StacksAddress, u64)],
 ) -> Result<(), String> {
     wait_for(timeout_secs, || {
-        let mut found_updates = HashSet::new();
+        let mut found_updates: HashSet<StacksAddress> = HashSet::new();
         let stackerdb_events = test_observer::get_stackerdb_chunks();
         for chunk in stackerdb_events
             .into_iter()
             .flat_map(|chunk| chunk.modified_slots)
         {
-            let message = SignerMessage::consensus_deserialize(&mut chunk.data.as_slice())
-                .expect("Failed to deserialize SignerMessage");
+            let Ok(message) = SignerMessage::consensus_deserialize(&mut chunk.data.as_slice())
+            else {
+                continue;
+            };
             let SignerMessage::StateMachineUpdate(update) = message else {
                 continue;
             };
             let Some((address, version)) = signer_addresses
                 .iter()
-                .find(|(addr, _)| chunk.verify(addr).unwrap())
+                .find(|(addr, _)| chunk.verify(addr).unwrap_or(false))
             else {
                 continue;
             };
-            match (version, &update.content) {
+            // Version sanity + expectation:
+            let negotiated = update
+                .active_signer_protocol_version
+                .min(update.local_supported_signer_protocol_version);
+            debug!("Found update from address {address} with version {version}. We expect version {negotiated}.");
+            if negotiated != *version {
+                continue;
+            }
+            match (negotiated, &update.content) {
                 (
                     0,
                     StateMachineUpdateContent::V0 {
@@ -1676,10 +1696,10 @@ pub fn wait_for_state_machine_update_by_miner_tenure_id(
                     },
                 ) => {
                     if tenure_id == expected_tenure_id {
-                        found_updates.insert(address);
+                        found_updates.insert(address.clone());
                     }
                 }
-                (_, _) => {}
+                (_, _) => continue,
             };
         }
         Ok(found_updates.len() > signer_addresses.len() * 7 / 10)
@@ -2732,7 +2752,8 @@ fn empty_tenure_delayed() {
         .collect();
     assert_eq!(signer_slot_ids.len(), num_signers);
 
-    // The miner's proposed block should get rejected by all the signers
+    // The miner's proposed block should get rejected by all the signers as the consensus hash does not match
+    // the active miner's tenure id
     let mut found_rejections = Vec::new();
     wait_for(short_timeout.as_secs(), || {
         for slot_id in signer_slot_ids.iter() {
@@ -2758,7 +2779,7 @@ fn empty_tenure_delayed() {
             })) = latest_msg
             {
                 assert_eq!(reason_code, RejectCode::SortitionViewMismatch);
-                assert_eq!(response_data.reject_reason, RejectReason::InvalidMiner);
+                assert!(matches!(response_data.reject_reason, RejectReason::ConsensusHashMismatch { .. }), "Unexpected reject reason: {}", response_data.reject_reason);
                 assert_eq!(metadata.server_version, VERSION_STRING.to_string());
                 found_rejections.push(*slot_id);
             } else {
@@ -7215,115 +7236,36 @@ fn verify_mempool_caches() {
     signer_test.shutdown();
 }
 
-/// Tests that signers are able to upgrade or downgrade their active protocol version numbers based on
-/// the majority of other signers current local supported version numbers
-#[test]
-#[ignore]
-fn rollover_signer_protocol_version() {
-    if env::var("BITCOIND_TEST") != Ok("1".into()) {
-        return;
-    }
-    let num_signers = 5;
-
-    let signer_test: SignerTest<SpawnedSigner> = SignerTest::new(num_signers, vec![]);
-    signer_test.boot_to_epoch_3();
-
-    let conf = signer_test.running_nodes.conf.clone();
-
-    let burnchain = conf.get_burnchain();
-    let sortdb = burnchain.open_sortition_db(true).unwrap();
-
-    let all_signers = signer_test.signer_test_pks();
-    info!(
-        "------------------------- Miner Tenure Starts and Mines Block N-------------------------"
-    );
+fn mine_burn_block_and_confirm_signer_rollover(
+    signer_test: &mut SignerTest<SpawnedSigner>,
+    expected_versions: &[(StacksAddress, u64)],
+) {
     test_observer::clear();
-    signer_test.mine_and_verify_confirmed_naka_block(Duration::from_secs(30), num_signers, true);
-
-    let tip = SortitionDB::get_canonical_burn_chain_tip(sortdb.conn()).unwrap();
-    let burn_consensus_hash = tip.consensus_hash;
-    let burn_height = tip.block_height;
-
-    info!("------------------------- Confirm Miner is the Active Miner in Update and All Signers Are Using Protocol Number {SUPPORTED_SIGNER_PROTOCOL_VERSION} -------------------------");
-    // Verify that signers first sent a bitcoin block update
-    wait_for_state_machine_update(
-        60,
-        &burn_consensus_hash,
-        burn_height,
-        None,
-        &signer_test.signer_addresses_versions(),
-    )
-    .expect("Timed out waiting for signers to send a state update for block N");
-
-    test_observer::clear();
-    let downgraded_version = SUPPORTED_SIGNER_PROTOCOL_VERSION.saturating_sub(1);
-    info!("------------------------- Downgrading Signer Versions to {downgraded_version} for 20 Percent of Signers -------------------------");
-    // Take a non blocking minority of signers (20%) and downgrade their version number
-    let pinned_signers: Vec<_> = all_signers
-        .iter()
-        .take(num_signers * 2 / 10)
-        .cloned()
-        .collect();
-    let pinned_signers_versions: HashMap<StacksPublicKey, u64> = pinned_signers
-        .iter()
-        .map(|signer| (signer.clone(), downgraded_version))
-        .collect();
-    TEST_PIN_SUPPORTED_SIGNER_PROTOCOL_VERSION.set(pinned_signers_versions);
-
-    info!("------------------------- Confirm Signers Still Manage to Sign a Stacks Block With Misaligned Version Numbers -------------------------");
-    signer_test.mine_and_verify_confirmed_naka_block(Duration::from_secs(30), num_signers, true);
-
-    let tip = SortitionDB::get_canonical_burn_chain_tip(sortdb.conn()).unwrap();
-    let burn_consensus_hash = tip.consensus_hash;
-    let burn_height = tip.block_height;
-    // Only one signer is downgraded so the active protocol version remains the same.
-    wait_for_state_machine_update(
-        60,
-        &burn_consensus_hash,
-        burn_height,
-        None,
-        &signer_test.signer_addresses_versions(),
-    )
-    .expect("Timed out waiting for signers to send their downgraded state update for block N+1");
-
-    test_observer::clear();
-    info!("------------------------- Confirm Signer Version Downgrades Fully Once 70 percent of Signers Downgrade -------------------------");
-    let pinned_signers: Vec<_> = all_signers
-        .iter()
-        .take(num_signers * 7 / 10)
-        .cloned()
-        .collect();
-    let pinned_signers_versions: HashMap<StacksPublicKey, u64> = pinned_signers
-        .iter()
-        .map(|signer| (signer.clone(), downgraded_version))
-        .collect();
-    TEST_PIN_SUPPORTED_SIGNER_PROTOCOL_VERSION.set(pinned_signers_versions);
-
-    // Not strictly necessary, but makes it easier to logic out if miner doesn't send a proposal until signers are on same page...
     TEST_MINE_SKIP.set(true);
-    info!("------------------------- Confirm Signers Sent Downgraded State Machine Updates -------------------------");
-    // Cannot use any built in functions that call mine_nakamoto_block since it expects signer updates matching the majority version and we are manually messing with these versions
+    info!("------------------------- Mine Bitcoin Block -------------------------");
     signer_test.mine_bitcoin_block();
-    let tip = SortitionDB::get_canonical_burn_chain_tip(sortdb.conn()).unwrap();
-    let burn_consensus_hash = tip.consensus_hash;
-    let burn_height = tip.block_height;
-    // Confirm ALL signers downgrade their supported version and then send a corresponding message in that version message
-    let downgraded_versions: Vec<_> = signer_test
-        .signer_addresses_versions()
-        .into_iter()
-        .map(|(address, _)| (address, downgraded_version))
-        .collect();
+    let tip = SortitionDB::get_canonical_burn_chain_tip(
+        signer_test
+            .running_nodes
+            .conf
+            .get_burnchain()
+            .open_sortition_db(true)
+            .unwrap()
+            .conn(),
+    )
+    .unwrap();
     wait_for_state_machine_update(
         60,
-        &burn_consensus_hash,
-        burn_height,
+        &tip.consensus_hash,
+        tip.block_height,
         None,
-        &downgraded_versions,
+        &expected_versions,
     )
-    .expect("Timed out waiting for signers to send their state update for block N+2");
-
+    .expect("Timed out waiting for signers to send their state updates after a bitcoin block");
     let info = signer_test.get_peer_info();
-    info!("------------------------- Confirm Signers Sign The Block After Complete Downgraded Version Number -------------------------");
+    info!(
+        "------------------------- Mine Tenure Change Stacks Transaction -------------------------"
+    );
     TEST_MINE_SKIP.set(false);
     let expected_miner = StacksPublicKey::from_private(
         &signer_test
@@ -7334,17 +7276,153 @@ fn rollover_signer_protocol_version() {
             .clone()
             .unwrap(),
     );
-    let block = wait_for_block_pushed_by_miner_key(60, info.stacks_tip_height + 1, &expected_miner)
-        .expect("Failed to mine block after downgraded version number.");
-    // Expect ALL signers even after downgrade to approve the proposed blocks
-    wait_for_block_acceptance_from_signers(30, &block.header.signer_signature_hash(), &all_signers)
-        .expect("Failed to confirm all signers accepted last block");
 
-    info!("------------------------- Reset All Signers to {SUPPORTED_SIGNER_PROTOCOL_VERSION} -------------------------");
-    TEST_PIN_SUPPORTED_SIGNER_PROTOCOL_VERSION.set(HashMap::new());
-    test_observer::clear();
-    info!("------------------------- Confirm Signers Sign The Block After Upgraded Version Number -------------------------");
-    signer_test.mine_and_verify_confirmed_naka_block(Duration::from_secs(30), num_signers, true);
+    let block = wait_for_block_pushed_by_miner_key(60, info.stacks_tip_height + 1, &expected_miner)
+        .expect("Failed to mine block after upgraded version number.");
+    wait_for_block_acceptance_from_signers(
+        30,
+        &block.header.signer_signature_hash(),
+        &signer_test.signer_test_pks(),
+    )
+    .expect("Failed to confirm all signers accepted block");
+}
+
+/// Tests that signers negotiate their **active** signer protocol version based on the
+/// **majority of locally supported** signer protocol versions.
+///
+/// Scenario (10 signers):
+/// 1) Baseline: all signers start on `SUPPORTED_SIGNER_PROTOCOL_VERSION` and can sign blocks.
+/// 2) Downgrade 30%: non-blocking minority downgraded. Stays at `SUPPORTED_SIGNER_PROTOCOL_VERSION`.
+/// 3) Downgrade 40%: blocking minority downgraded. Downgrades to `SUPPORTED_SIGNER_PROTOCOL_VERSION - 1`.
+/// 4) Upgrade 70%: majority upgraded. Upgrades back to `SUPPORTED_SIGNER_PROTOCOL_VERSION`.
+#[test]
+#[ignore]
+fn downgrade_signer_protocol_version() {
+    if env::var("BITCOIND_TEST") != Ok("1".into()) {
+        return;
+    }
+    // To make the math easier, use 10 signers
+    let num_signers: usize = 10;
+
+    // Integer percent helper (e.g. 30 => 3 when num_signers=10)
+    fn pct(num_signers: usize, percent: usize) -> usize {
+        (num_signers * percent) / 100
+    }
+
+    // Boot with the most recent supported version
+    let mut signer_test: SignerTest<SpawnedSigner> = SignerTest::new(num_signers, vec![]);
+    signer_test.boot_to_epoch_3();
+
+    info!(
+        "------------------------- Confirm miner tenure continues and all signers use protocol {SUPPORTED_SIGNER_PROTOCOL_VERSION} -------------------------"
+    );
+    let versions = signer_test.signer_addresses_versions();
+    mine_burn_block_and_confirm_signer_rollover(&mut signer_test, &versions);
+
+    let downgraded_version = SUPPORTED_SIGNER_PROTOCOL_VERSION.saturating_sub(1);
+    let downgraded_versions = signer_test
+        .signer_addresses_versions()
+        .into_iter()
+        .map(|(addr, _)| (addr, downgraded_version))
+        .collect::<Vec<_>>();
+    info!("------------------------- Downgrading Signer Versions to {downgraded_version} for 30 Percent of Signers -------------------------");
+    // Since only 30 percent downgraded. Active protocol version should stay the same
+    signer_test
+        .restart_first_n_signers_with_supported_version(pct(num_signers, 30), downgraded_version);
+    let versions = signer_test.signer_addresses_versions();
+    mine_burn_block_and_confirm_signer_rollover(&mut signer_test, &versions);
+    info!("------------------------- Downgrading Signer Versions to {downgraded_version} for 40 Percent of Signers -------------------------");
+    // A blocking minority has downgraded, so expect everyone to negotiate to downgraded_version.
+    signer_test
+        .restart_first_n_signers_with_supported_version(pct(num_signers, 40), downgraded_version);
+    mine_burn_block_and_confirm_signer_rollover(&mut signer_test, &downgraded_versions);
+
+    info!("------------------------- Upgrade Signer Versions to {SUPPORTED_SIGNER_PROTOCOL_VERSION} for 70 Percent of Signers -------------------------");
+    signer_test.restart_first_n_signers_with_supported_version(
+        pct(num_signers, 70),
+        SUPPORTED_SIGNER_PROTOCOL_VERSION,
+    );
+    let versions = signer_test.signer_addresses_versions();
+    // A majority of signers have upgraded, expect everyone to negotiate back to the supported version.
+    mine_burn_block_and_confirm_signer_rollover(&mut signer_test, &versions);
+    signer_test.shutdown();
+}
+
+/// Tests the latest activation rollover behavior across different activation
+/// percentages, ensuring that block signing continues as old signers can still communicate with new signers.
+///
+/// Scenario (10 signers):
+/// 1) 30% activated: non-blocking minority activated.
+/// 2) 40% activated: blocking minority activated .
+/// 3) 60% activated: blocking minority unactivated.
+/// 4) 70% activated: non-blocking minority unactivated.
+/// 5) 100% activated: all signers activated.
+#[test]
+#[ignore]
+fn rollover_signer_protocol_version() {
+    if env::var("BITCOIND_TEST") != Ok("1".into()) {
+        return;
+    }
+
+    // To make the math easier, use 10 signers
+    let num_signers: usize = 10;
+
+    // Integer percent helper (e.g. 30 => 3 when num_signers=10)
+    fn pct(num_signers: usize, percent: usize) -> usize {
+        (num_signers * percent) / 100
+    }
+
+    let old_version = SUPPORTED_SIGNER_PROTOCOL_VERSION.saturating_sub(1);
+    // Start with 30% of signers activated
+    let mut signer_test: SignerTest<SpawnedSigner> = SignerTest::new_with_config_modifications(
+        num_signers,
+        vec![],
+        |signer_config| {
+            // 30% of 10 is 3 signers
+            let activated_signers = pct(num_signers, 30) as u16;
+
+            // signer_nmb in [0, num_signers-1]
+            let signer_nmb = signer_config.endpoint.port() % num_signers as u16;
+
+            let signer_version = if signer_nmb < activated_signers {
+                SUPPORTED_SIGNER_PROTOCOL_VERSION
+            } else {
+                old_version
+            };
+
+            signer_config.supported_signer_protocol_version = signer_version;
+        },
+        |_| {},
+        None,
+        None,
+    );
+
+    signer_test.boot_to_epoch_3();
+
+    let old_versions: Vec<_> = signer_test
+        .signer_addresses_versions()
+        .into_iter()
+        .map(|(address, _)| (address, old_version))
+        .collect();
+
+    mine_burn_block_and_confirm_signer_rollover(&mut signer_test, &old_versions);
+
+    for percent in [40, 60, 70] {
+        info!("------------------------- Upgrade to {SUPPORTED_SIGNER_PROTOCOL_VERSION} for {percent}% of signers -------------------------");
+        signer_test.restart_first_n_signers_with_supported_version(
+            pct(num_signers, percent),
+            SUPPORTED_SIGNER_PROTOCOL_VERSION,
+        );
+        mine_burn_block_and_confirm_signer_rollover(&mut signer_test, &old_versions);
+    }
+
+    info!("------------------------- Upgrade to {SUPPORTED_SIGNER_PROTOCOL_VERSION} for 100% of signers -------------------------");
+    signer_test.restart_first_n_signers_with_supported_version(
+        num_signers,
+        SUPPORTED_SIGNER_PROTOCOL_VERSION,
+    );
+    let new_versions = signer_test.signer_addresses_versions();
+    mine_burn_block_and_confirm_signer_rollover(&mut signer_test, &new_versions);
 
     signer_test.shutdown();
 }
@@ -7920,19 +7998,37 @@ fn signers_do_not_commit_unless_threshold_precommitted() {
     let ignore_signers: Vec<_> = ignore_slice.to_vec();
     let pre_commit_signers: Vec<_> = pre_commit_slice.to_vec();
     TEST_IGNORE_ALL_BLOCK_PROPOSALS.set(ignore_signers);
+    // Pause the miner proposing blocks because we need to make sure that the global signer state updates before we evaluate the proposal to ensure
+    // it is isn't immediately rejected under global signer state conditions
+    TEST_BROADCAST_PROPOSAL_STALL.set(vec![miner_pk.clone()]);
     test_observer::clear();
     let blocks_before = test_observer::get_mined_nakamoto_blocks().len();
     let height_before = signer_test.get_peer_info().stacks_tip_height;
+    info!("------------------------- Start Tenure A -------------------------");
     next_block_and(
         &mut signer_test.running_nodes.btc_regtest_controller,
         30,
         || Ok(test_observer::get_mined_nakamoto_blocks().len() > blocks_before),
     )
     .unwrap();
+    let peer_after = signer_test.get_peer_info();
+    wait_for_state_machine_update(
+        30,
+        &peer_after.pox_consensus,
+        peer_after.burn_block_height,
+        None,
+        &signer_test.signer_addresses_versions(),
+    )
+    .expect("Failed to update to tenure A");
 
+    TEST_BROADCAST_PROPOSAL_STALL.set(vec![]);
+
+    info!("------------------------- Wait for block proposal N -------------------------");
     let proposal = wait_for_block_proposal(30, height_before + 1, &miner_pk)
         .expect("Timed out waiting for block proposal");
     let hash = proposal.header.signer_signature_hash();
+
+    info!("------------------------- Wait for pre-commits for N ({hash})-------------------------");
     wait_for_block_pre_commits_from_signers(30, &hash, &pre_commit_signers)
         .expect("Timed out waiting for pre-commits");
     assert!(